fixes #30; no inline formatting in headings (was: backticks in headings not interpreted)
