+++ test/regressions_test.exs
@@ -109,4 +109,14 @@ defmodule RegressionsTest do
                      &quot;Hello &amp;lt;world&amp;gt;&quot;</code></pre>
                      """
   end
+
+  @heading_inline_render """
+  # Hello _World_
+  """
+  test "Issue https://github.com/pragdave/earmark/issues/30" do
+    result = Earmark.to_html @heading_inline_render
+    assert result == """
+                     <h1>Hello <em>World</em></h1>
+                     """
+  end
 end
